@@ -1,9 +1,8 @@
-import { ColorModeScript } from '@chakra-ui/react';
 import axios from 'axios';
 import dayjs from 'dayjs';
 import 'dayjs/locale/pl';
 import * as React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import { App } from './App';
 import reportWebVitals from './reportWebVitals';
@@ -13,13 +12,8 @@ import * as serviceWorker from './serviceWorker';
 axios.defaults.baseURL = 'https://testnginx.bieda.it/';
 dayjs.locale('pl');
 
-ReactDOM.render(
-  <React.StrictMode>
-    <ColorModeScript />
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+const root = createRoot(document.getElementById('root') as any);
+root.render(<App />);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
