@@ -1,10 +1,10 @@
-import { ArrowRightIcon } from '@chakra-ui/icons';
 import { Box, Button, Flex, Heading, Input, Spacer, Text, VStack } from '@chakra-ui/react';
+import { Laptop } from '../components/landing/Laptop';
 
 export const LandingPage = () => {
   return (
-    <Flex p={4} m={0} h={['100vh']} direction={'row'} overflow={'hidden'}>
-      <Flex h={'100%'} direction={'column'} w={'50%'}>
+    <Flex p={4} m={0} h={['100vh']} direction={['column', 'row']} overflow={'hidden'}>
+      <Flex h={['50%', '100%']} direction={'column'} w={['100%', '50%']} order={[2, 1]}>
         <Spacer />
         <Heading textAlign={'center'} fontSize={'5xl'}>
           Skończ z tym
@@ -17,53 +17,8 @@ export const LandingPage = () => {
         </Button>
         <Spacer />
       </Flex>
-      <Flex h={'100%'} direction={'column'} w={'50%'} alignItems={'center'}>
-        <Spacer />
-        <Flex
-          h={'80%'}
-          w={'60%'}
-          backgroundColor={'gray.700'}
-          p={2}
-          borderRadius={'10'}
-          shadow={'dark-lg'}
-          direction={'column'}
-          gap={2}
-        >
-          <Flex
-            backgroundColor={'gray.900'}
-            w={'100%'}
-            h={'100%'}
-            borderRadius={'10'}
-            direction={'column'}
-            p={3}
-            alignItems={'baseline'}
-            gap={1}
-          >
-            <VStack spacing={0} alignItems={'baseline'}>
-              <Text ml={1} fontSize={'xs'}>
-                Jerzy
-              </Text>
-              <Box backgroundColor={'gray.700'} p={2} borderRadius={10} maxW={'70%'}>
-                <Text>Kiedy kolos z matematyki?</Text>
-              </Box>
-            </VStack>
-            <VStack spacing={0} alignItems={'baseline'}>
-              <Text ml={1} fontSize={'xs'}>
-                Marzena
-              </Text>
-              <Box backgroundColor={'gray.700'} p={2} borderRadius={10} maxW={'70%'}>
-                <Text>Do kiedy jest termin wysyłania projektu z analizy matematycznej?</Text>
-              </Box>
-            </VStack>
-          </Flex>
-          <Flex w={'100%'} borderRadius={'10'} direction={'row'} alignItems={'baseline'} gap={1}>
-            <Input placeholder='Wiadomość...' backgroundColor={'gray.800'} disabled mr={1} />
-            <Button colorScheme={'blue'} disabled>
-              Wyślij
-            </Button>
-          </Flex>
-        </Flex>
-        <Spacer />
+      <Flex h={['50%', '100%']} direction={'column'} w={['100%', '50%']} order={[1, 2]}>
+        <Laptop />
       </Flex>
     </Flex>
   );
