@@ -1,7 +1,7 @@
 import create from 'zustand';
 import { devtools, persist } from 'zustand/middleware';
 import { createAuthSlice } from './authSlice';
-import axios from "axios";
+import axios from 'axios';
 
 const store = (set: any) => ({
   ...createAuthSlice(set),
